+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,0 +1,17 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AdminController {
+    @GetMapping("/admin")
+    public String openAdminPage() {
+        return "/admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String openAdminReservationPage() {
+        return "/admin/reservation-legacy";
+    }
+}
