+++ src/main/java/roomescape/controller/ReservationController.java
@@ -4,37 +4,25 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
 import roomescape.dto.ReservationCreateDto;
 
-@Controller
-public class RoomEscapeController {
+@RestController
+public class ReservationController {
     private List<Reservation> reservations = new ArrayList<>();
     private AtomicLong index = new AtomicLong(1);
-
-    @GetMapping("/admin")
-    public String openAdminPage() {
-        return "/admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String openReservationPage() {
-        return "/admin/reservation-legacy";
-    }
 
     @GetMapping("/reservations")
     public ResponseEntity<List<Reservation>> readReservations() {
         return ResponseEntity.ok(reservations);
     }
 
-    @ResponseBody
     @PostMapping("/reservations")
     public Reservation createReservation(@RequestBody ReservationCreateDto dto) {
         Reservation reservation = new Reservation(index.getAndIncrement(), dto.getName(), dto.getDate(), dto.getTime());
